requirements.txt
# Dependencias Principales
pandas
pandas-ta
python-binance
python-dotenv
websocket-client
tqdm

# Dependencias de Análisis y Reportes
quantstats
IPython

# === NUEVA DEPENDENCIA PARA FASE 2 ===
# Análisis de Ondas de Elliott
python-taew

# Nota: python-taew automatiza la identificación de ondas impulsivas y correctivas
# según la Teoría de las Ondas de Elliott, permitiendo análisis estructural avanzado
# del mercado para nuestro nuevo paradigma de trading.